+++ services/el_clients/geth/specular/rollup/services/indexer/indexer.go
@@ -3,9 +3,9 @@ package indexer
 import (
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/rpc"
-	"github.com/specularl2/specular/clients/geth/specular/proof"
-	"github.com/specularl2/specular/clients/geth/specular/rollup/client"
-	"github.com/specularl2/specular/clients/geth/specular/rollup/services"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/proof"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup/client"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup/services"
 )
 
 type Indexer struct{ *services.BaseService }
