+++ services/el_clients/geth/specular/rollup/services/sequencer/sequencer.go
@@ -10,13 +10,13 @@ import (
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/rpc"
-	"github.com/specularl2/specular/clients/geth/specular/bindings"
-	"github.com/specularl2/specular/clients/geth/specular/proof"
-	"github.com/specularl2/specular/clients/geth/specular/rollup/client"
-	"github.com/specularl2/specular/clients/geth/specular/rollup/services"
-	rollupTypes "github.com/specularl2/specular/clients/geth/specular/rollup/types"
-	"github.com/specularl2/specular/clients/geth/specular/rollup/utils/fmt"
-	"github.com/specularl2/specular/clients/geth/specular/rollup/utils/log"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/bindings"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/proof"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup/client"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup/services"
+	rollupTypes "github.com/specularl2/specular/services/el_clients/geth/specular/rollup/types"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup/utils/fmt"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup/utils/log"
 )
 
 const timeInterval = 3 * time.Second
