+++ services/el_clients/geth/specular/rollup/utils/broker.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/ethereum/go-ethereum/event"
-	"github.com/specularl2/specular/clients/geth/specular/rollup/utils/log"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup/utils/log"
 )
 
 type Broker[T any] struct {
